Move SoundEntry builder to dedicated subpackage
